+++ shroud/generate.py
@@ -120,7 +120,6 @@ class VerifyAttrs(object):
             if attr not in [
                 "api",          # arguments to pass to C wrapper.
                 "allocatable",  # return a Fortran ALLOCATABLE
-                "cdesc",
                 "deref",  # How to dereference pointer
                 "dimension",
                 "free_pattern",
@@ -354,9 +353,9 @@ class VerifyAttrs(object):
         api = attrs["api"]
         if api is None:
             pass
-        elif api not in ["capi", "buf", "cfi"]:
+        elif api not in ["capi", "buf", "cdesc", "cfi"]:
                 raise RuntimeError(
-                    "'api' attribute must 'capi', 'buf', or 'cfi'"
+                    "'api' attribute must 'capi', 'buf', 'cdesc' or 'cfi'"
                 )
         else:
             meta["api"] = api
@@ -461,7 +460,6 @@ class VerifyAttrs(object):
                 "assumedtype",
                 "blanknull",   # Treat blank string as NULL pointer.
                 "capsule",
-                "cdesc",
                 "charlen",   # Assumed length of intent(out) char *.
                 "external",
                 "deref",
@@ -1794,10 +1792,7 @@ class GenFunctions(object):
             meta = declarator.metaattrs
             if meta["api"]:
                 # API explicitly set by user.
-                continue
-            elif attrs["cdesc"]:
-                # User requested cdesc.
-                has_buf_arg = "cdesc"
+                pass
             elif arg_typemap.sgroup == "string":
                 if meta["deref"] in ["allocatable", "pointer", "copy"]:
                     has_buf_arg = "cdesc"